+++ cmake/cpack_settings.cmake
@@ -12,9 +12,9 @@ set(CPACK_PACKAGE_VERSION ${OE_VERSION})
 set(CPACK_PACKAGING_INSTALL_PREFIX ${CMAKE_INSTALL_PREFIX})
 set(CPACK_DEBIAN_PACKAGE_DEPENDS "libsgx-enclave-common (>=2.3.100.46354-1), libsgx-enclave-common-dev (>=2.3.100.0-1), libsgx-dcap-ql (>=1.0.100.46460-1.0), libsgx-dcap-ql-dev (>=1.0.100.46460-1.0), pkg-config")
 
-# Cpack variables for quote verification component
-set(CPACK_DEBIAN_OEQUOTE_PACKAGE_NAME "open-enclave-quote")
-set(CPACK_DEBIAN_OEQUOTE_FILE_NAME DEB-DEFAULT)
-set(CPACK_COMPONENT_OEQUOTE_DESCRIPTION "Open Enclave Quote Verification Library")
-set(CPACK_DEBIAN_OEQUOTE_PACKAGE_DEPENDS "pkg-config")
+# Cpack variables for host verification component
+set(CPACK_DEBIAN_OEHOSTVERIFY_PACKAGE_NAME "open-enclave-hostverify")
+set(CPACK_DEBIAN_OEHOSTVERIFY_FILE_NAME DEB-DEFAULT)
+set(CPACK_COMPONENT_OEHOSTVERIFY_DESCRIPTION "Open Enclave Report Verification Host Library")
+set(CPACK_DEBIAN_OEHOSTVERIFY_PACKAGE_DEPENDS "pkg-config")
 include(CPack)